move remove_color_codes fn and tests to proper files
