+++ tests/test_utils.py
@@ -12,7 +12,6 @@ from autogpt.utils import (
     get_current_git_branch,
     get_latest_bulletin,
     readable_file_size,
-    remove_color_codes,
     validate_yaml_file,
 )
 from tests.utils import skip_in_ci
@@ -195,48 +194,6 @@ def test_get_latest_bulletin_new_bulletin_same_as_old_bulletin():
         assert bulletin == "Test bulletin"
 
     os.remove("CURRENT_BULLETIN.md")
-
-
-@pytest.mark.parametrize("raw_text, clean_text", [
-    ( "COMMAND = \x1b[36mbrowse_website\x1b[0m  ARGUMENTS = \x1b[36m{'url': 'https://www.google.com', question': 'What is the capital of France?'}\x1b[0m", 
-     "COMMAND = browse_website  ARGUMENTS = {'url': 'https://www.google.com', 'question': 'What is the capital of France?'}"), 
-     ("{'Schaue dir meine Projekte auf github () an, als auch meine Webseiten': https://github.com/Significant-Gravitas/Auto-GPT, https://discord.gg/autogpt und https://twitter.com/SigGravitas'}",
-     "{'Schaue dir meine Projekte auf github () an, als auch meine Webseiten': 'https://github.com/Significant-Gravitas/Auto-GPT, https://discord.gg/autogpt und https://twitter.com/SigGravitas'}"),
-     ("", ""),
-     ("hello", "hello"),
-     ("hello\x1B[31m world", "hello world"),
-     ("\x1B[36mHello,\x1B[32m World!", "Hello, World!"),
-     ("\x1B[1m\x1B[31mError:\x1B[0m\x1B[31m file not found", "Error: file not found"),
-)
-def test_remove_color_codes(raw_text, clean_text):
-    assert remove_color_codes(raw_text) == clean_text
-    assert (
-        remove_color_codes(
-            "COMMAND = \x1b[36mbrowse_website\x1b[0m  ARGUMENTS = \x1b[36m{'url': 'https://www.google.com', "
-            "'question': 'What is the capital of France?'}\x1b[0m"
-        )
-        == "COMMAND = browse_website  ARGUMENTS = {'url': 'https://www.google.com', 'question': 'What is the capital "
-        "of France?'}"
-    )
-    assert (
-        remove_color_codes(
-            "{'Schaue dir meine Projekte auf github () an, als auch meine Webseiten': "
-            "'https://github.com/Significant-Gravitas/Auto-GPT, https://discord.gg/autogpt und "
-            "https://twitter.com/SigGravitas'}"
-        )
-        == "{'Schaue dir meine Projekte auf github () an, als auch meine Webseiten': "
-        "'https://github.com/Significant-Gravitas/Auto-GPT, https://discord.gg/autogpt und "
-        "https://twitter.com/SigGravitas'}"
-    )
-    assert remove_color_codes("") == ""
-    assert remove_color_codes("hello") == "hello"
-    assert remove_color_codes("hello\x1B[31m world") == "hello world"
-    assert remove_color_codes("\x1B[36mHello,\x1B[32m World!") == "Hello, World!"
-    assert (
-        remove_color_codes("\x1B[1m\x1B[31mError:\x1B[0m\x1B[31m file not found")
-        == "Error: file not found"
-    )
-    assert remove_color_codes({"I": "am a dict"}) == f"{{'I': 'am a dict'}}"
 
 
 if __name__ == "__main__":
